Fixed inability for CCS project to compile by replacing admin files
Because:
1. This was halting progress for software development for the ETCM

The CCS project empty_driverlib_project successfully compiles. Thus, the BOLT_IV_ETCM folders ".settings", "CPU1_RAM", "device", "targetConfigs" and the files ".ccsproject", ".cproject", ".project", "2837xD_FLAHS_Ink_cpu1.cmd" and "2837xD_RAM_Ink_cpu1.cmd" where replace with those from empty_driberlib_project. The original ones were first deleted and then the new ones were copied and pasted.

+++ device/driverlib/eqep.c
@@ -5,10 +5,10 @@
 // TITLE:  C28x eQEP driver.
 //
 //###########################################################################
-// $TI Release: F2837xD Support Library v3.05.00.00 $
-// $Release Date: Tue Jun 26 03:15:23 CDT 2018 $
+// $TI Release: F2837xD Support Library v3.07.00.00 $
+// $Release Date: Sun Sep 29 07:34:54 CDT 2019 $
 // $Copyright:
-// Copyright (C) 2013-2018 Texas Instruments Incorporated - http://www.ti.com/
+// Copyright (C) 2013-2019 Texas Instruments Incorporated - http://www.ti.com/
 //
 // Redistribution and use in source and binary forms, with or without 
 // modification, are permitted provided that the following conditions 
@@ -67,8 +67,8 @@ EQEP_setCompareConfig(uint32_t base, uint16_t config, uint32_t compareValue,
     //
     // Set the shadow register settings and pulse width.
     //
-    regValue = (config & (EQEP_QPOSCTL_PCSHDW | EQEP_QPOSCTL_PCLOAD)) |
-               (cycles - 1U);
+    regValue = (config & (uint16_t)(EQEP_QPOSCTL_PCSHDW |
+                  EQEP_QPOSCTL_PCLOAD)) | (cycles - 1U);
 
     HWREGH(base + EQEP_O_QPOSCTL) = (HWREGH(base + EQEP_O_QPOSCTL) &
                                      ~(EQEP_QPOSCTL_PCSPW_M |
@@ -78,7 +78,7 @@ EQEP_setCompareConfig(uint32_t base, uint16_t config, uint32_t compareValue,
     //
     // Set position compare sync-output mode.
     //
-    regValue = config & (EQEP_QDECCTL_SOEN | EQEP_QDECCTL_SPSEL);
+    regValue = config & (uint16_t)(EQEP_QDECCTL_SOEN | EQEP_QDECCTL_SPSEL);
 
     HWREGH(base + EQEP_O_QDECCTL) = (HWREGH(base + EQEP_O_QDECCTL) &
                                      ~(EQEP_QDECCTL_SOEN |
